requirements.txt
anthropic>=0.25.0
crawl4ai>=0.3.0
httpx>=0.27.0
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0